putting gaussian prior back where it lives
