+++ autofit/mapper/prior/gaussian.py
@@ -1,5 +1,1 @@
-from autofit.messages.normal import NormalMessage
 
-
-class GaussianPrior(NormalMessage):
-    """A prior with a gaussian distribution"""
