+++ autofit/mapper/prior_model/abstract.py
@@ -16,8 +16,7 @@ from autofit import exc
 from autofit.mapper import model
 from autofit.mapper.model import AbstractModel, frozen_cache
 from autofit.mapper.prior.deferred import DeferredArgument
-from autofit.mapper.prior.gaussian import GaussianPrior
-from autofit.mapper.prior.prior import Prior, Limits
+from autofit.mapper.prior.prior import Prior, Limits, GaussianPrior
 from autofit.mapper.prior.tuple_prior import TuplePrior
 from autofit.mapper.prior.width_modifier import WidthModifier
 from autofit.mapper.prior_model.attribute_pair import DeferredNameValue
